requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==5.3.1
    # via kombu
apiflask==2.4.0
    # via -r requirements.in
apispec==6.8.2
    # via apiflask
argon2-cffi==25.1.0
    # via minio
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
attrs==25.3.0
    # via
    #   cattrs
    #   requests-cache
billiard==4.2.1
    # via celery
blinker==1.9.0
    # via flask
cattrs==25.1.1
    # via requests-cache
celery==5.5.3
    # via -r requirements.in
certifi==2025.8.3
    # via
    #   minio
    #   requests
cffi==1.17.1
    # via argon2-cffi-bindings
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
    #   rich-click
    #   roc-validator
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorlog==6.9.0
    # via roc-validator
enum-tools==0.12.0
    # via roc-validator
flask==3.0.3
    # via
    #   -r requirements.in
    #   apiflask
    #   flask-httpauth
    #   flask-marshmallow
flask-httpauth==4.8.0
    # via apiflask
flask-marshmallow==1.3.0
    # via apiflask
html5rdf==1.2.1
    # via rdflib
idna==3.10
    # via
    #   requests
    #   url-normalize
importlib-metadata==8.7.0
    # via pyshacl
inquirerpy==0.3.4
    # via roc-validator
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
kombu==5.5.4
    # via celery
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
marshmallow==4.0.0
    # via
    #   apiflask
    #   flask-marshmallow
    #   webargs
mdurl==0.1.2
    # via markdown-it-py
minio==7.2.16
    # via -r requirements.in
owlrl==7.1.4
    # via pyshacl
packaging==25.0
    # via
    #   apispec
    #   kombu
    #   pyshacl
    #   webargs
pfzy==0.3.4
    # via inquirerpy
platformdirs==4.3.8
    # via requests-cache
prettytable==3.16.0
    # via pyshacl
prompt-toolkit==3.0.51
    # via
    #   click-repl
    #   inquirerpy
pycparser==2.22
    # via cffi
pycryptodome==3.23.0
    # via minio
pygments==2.19.2
    # via
    #   enum-tools
    #   rich
pyparsing==3.2.3
    # via rdflib
pyshacl==0.30.1
    # via roc-validator
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.1.1
    # via -r requirements.in
rdflib[html]==7.1.4
    # via
    #   owlrl
    #   pyshacl
    #   roc-validator
redis==5.2.0
    # via -r requirements.in
requests==2.32.4
    # via
    #   -r requirements.in
    #   requests-cache
    #   roc-validator
requests-cache==1.2.1
    # via roc-validator
rich==13.9.4
    # via
    #   rich-click
    #   roc-validator
rich-click==1.8.9
    # via roc-validator
roc-validator==0.7.3
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
toml==0.10.2
    # via roc-validator
typing-extensions==4.14.1
    # via
    #   cattrs
    #   enum-tools
    #   minio
    #   rich-click
tzdata==2025.2
    # via kombu
url-normalize==2.2.1
    # via requests-cache
urllib3==2.5.0
    # via
    #   minio
    #   requests
    #   requests-cache
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   prettytable
    #   prompt-toolkit
webargs==8.7.0
    # via apiflask
werkzeug==3.0.6
    # via
    #   -r requirements.in
    #   flask
zipp==3.23.0
    # via importlib-metadata
